+++ src/main/java/com/woowacamp/soolsool/core/liquor/code/LiquorErrorCode.java
@@ -1,4 +1,4 @@
-package com.woowacamp.soolsool.core.liquor.exception;
+package com.woowacamp.soolsool.core.liquor.code;
 
 import static org.springframework.http.HttpStatus.NOT_FOUND;
