+++ src/main/java/com/woowacamp/soolsool/core/liquor/domain/vo/LiquorImageUrl.java
@@ -1,6 +1,6 @@
 package com.woowacamp.soolsool.core.liquor.domain.vo;
 
-import com.woowacamp.soolsool.core.liquor.exception.LiquorErrorCode;
+import com.woowacamp.soolsool.core.liquor.code.LiquorErrorCode;
 import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
