+++ src/test/java/com/woowacamp/soolsool/integration/repository/LiquorRegionRepositoryTest.java
@@ -3,8 +3,8 @@ package com.woowacamp.soolsool.integration.repository;
 import static com.woowacamp.soolsool.core.liquor.domain.vo.LiquorRegionType.GYEONGGI_DO;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.woowacamp.soolsool.core.liquor.code.LiquorErrorCode;
 import com.woowacamp.soolsool.core.liquor.domain.LiquorRegion;
-import com.woowacamp.soolsool.core.liquor.exception.LiquorErrorCode;
 import com.woowacamp.soolsool.core.liquor.repository.LiquorRegionRepository;
 import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import org.junit.jupiter.api.DisplayName;
