+++ src/test/java/com/woowacamp/soolsool/integration/repository/LiquorStatusTypeRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.woowacamp.soolsool.integration.repository;
 
+import static com.woowacamp.soolsool.core.liquor.code.LiquorErrorCode.NOT_LIQUOR_STATUS_FOUND;
 import static com.woowacamp.soolsool.core.liquor.domain.vo.LiquorStatusType.ON_SALE;
-import static com.woowacamp.soolsool.core.liquor.exception.LiquorErrorCode.NOT_LIQUOR_STATUS_FOUND;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.woowacamp.soolsool.core.liquor.domain.LiquorStatus;
