+++ src/test/java/com/woowacamp/soolsool/integration/repository/LiquorTypeRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.woowacamp.soolsool.integration.repository;
 
+import static com.woowacamp.soolsool.core.liquor.code.LiquorErrorCode.NOT_LIQUOR_BREW_FOUND;
 import static com.woowacamp.soolsool.core.liquor.domain.vo.LiquorBrewType.BERRY;
-import static com.woowacamp.soolsool.core.liquor.exception.LiquorErrorCode.NOT_LIQUOR_BREW_FOUND;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.woowacamp.soolsool.core.liquor.domain.LiquorBrew;
