+++ src/test/java/com/woowacamp/soolsool/integration/service/LiquorServiceTest.java
@@ -1,6 +1,6 @@
 package com.woowacamp.soolsool.integration.service;
 
-import static com.woowacamp.soolsool.core.liquor.exception.LiquorErrorCode.NOT_LIQUOR_FOUND;
+import static com.woowacamp.soolsool.core.liquor.code.LiquorErrorCode.NOT_LIQUOR_FOUND;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.junit.jupiter.api.Assertions.assertAll;
